requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
reportlab>=4.0.0
numpy>=1.24.0
opencv-python>=4.5.4
tkinterdnd2>=0.3.0
pillow-heif>=0.16.0
